+++ CHANGELOG.md
@@ -13,6 +13,7 @@ TESTS:
 - Added etcd v2 provider tests.
 - Add test for degraded cluster (TestDegradedCluster).
 - Add more tests for etcd v3 viper provider.
+- Add tests and correct panics for invalid topology config from viper.
 
 ## v2.0.1